requirements.txt
# Core ML & Data Science
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# MLOps & Experiment Tracking
mlflow>=2.4.0
wandb>=0.15.0
dvc>=3.0.0

# API & Web Framework
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring & Observability
prometheus-client>=0.16.0
grafana-api>=1.0.3
python-dotenv>=1.0.0

# Testing & Code Quality
pytest>=7.3.1
pytest-cov>=4.0.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0

# Infrastructure & Cloud
boto3>=1.26.0
google-cloud-storage>=2.9.0
kubernetes>=26.1.0
docker>=6.1.0 